requirements.txt
Django==1.9.5
PyJWT==1.4.0
requests==2.9.1